wip: decomposition of settings
