+++ src/pages/Account/EmailSetting.js
@@ -1,9 +1,12 @@
 import React from 'react'
 import gql from 'graphql-tag'
+import { compose } from 'recompose'
 import { graphql } from 'react-apollo'
 import { store } from '../../index'
 import { showNotification } from '../../actions/rootActions'
 import EditableInputSetting from './EditableInputSetting'
+import * as actions from '../../actions/types'
+import { connect } from 'react-redux'
 
 const TAKEN_MSG = 'Email has already been taken'
 
@@ -26,6 +29,7 @@ const validateEmail = email => {
 }
 
 const EmailSetting = ({ email, dispatchNewEmail, changeEmail }) => {
+  console.log(email)
   return (
     <EditableInputSetting
       label='Email'
@@ -54,6 +58,24 @@ const EmailSetting = ({ email, dispatchNewEmail, changeEmail }) => {
   )
 }
 
-export default graphql(CHANGE_EMAIL_MUTATION, { name: 'changeEmail' })(
-  EmailSetting
+const mapStateToProps = ({ user: { data: { email } = {} } }) => ({
+  email
+})
+
+const mapDispatchToProps = dispatch => ({
+  dispatchNewEmail: email =>
+    dispatch({
+      type: actions.USER_EMAIL_CHANGE,
+      email
+    })
+})
+
+const enhance = compose(
+  connect(
+    mapStateToProps,
+    mapDispatchToProps
+  ),
+  graphql(CHANGE_EMAIL_MUTATION, { name: 'changeEmail' })
 )
+
+export default enhance(EmailSetting)
